+++ lib/injection_container.dart
@@ -1,11 +1,9 @@
-import 'dart:io';
-
 import 'package:get_it/get_it.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'data/managers/custom_cache_manager.dart';
 import 'data/managers/local_store_manager.dart';
-import 'data/repos/api_repo.dart';
+import 'data/managers/shared_preference_manager.dart';
 
 final sl = GetIt.instance;
 
@@ -14,7 +12,7 @@ Future<void> initializeDependencies() async {
   final sharedPreferences = await SharedPreferences.getInstance();
   sl.registerSingleton<SharedPreferences>(sharedPreferences);
 
-  // Managers
+  // Register
   sl.registerSingleton<LocalStorageManager>(
     SharedPreferenceManager(sl<SharedPreferences>()),
   );