+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Create package
+- Added example
+- Updated documentation